refactor: Centralize Cat schema definition
Refactored the codebase to define the Cat schema (Cat, CatId, CatIdFromString) in a central file `packages/domain/src/cats.ts`.

Updated `CatsApi.ts`, `CatsRepository.ts`, and corresponding test files (`CatsRepository.test.ts`, `CatsApi.integration.test.ts`) to import the Cat schema from this new location.

This change improves modularity and clarity of the domain model.
