+++ packages/domain/src/CatsApi.ts
@@ -1,22 +1,11 @@
-import { HttpApi, HttpApiEndpoint, HttpApiGroup } from "@effect/platform"
-import { Schema } from "effect"
+import { HttpApi, HttpApiEndpoint, HttpApiGroup } from "@effect/platform";
+import { Schema } from "effect";
+import { Cat, CatIdFromString } from "./cats.js"; // Import Cat and CatIdFromString
 
-export const CatId = Schema.Number.pipe(Schema.brand("CatId"))
-export type CatId = typeof CatId.Type
-
-export const CatIdFromString = Schema.NumberFromString.pipe(
-  Schema.compose(CatId)
-)
-
-export class Cat extends Schema.Class<Cat>("Cat")({
-  id: CatId,
-  name: Schema.NonEmptyTrimmedString,
-  breed: Schema.NonEmptyTrimmedString,
-  age: Schema.Number
-}) {}
+// CatId and Cat class are now imported from ./cats.ts
 
 export class CatNotFound extends Schema.TaggedError<CatNotFound>()("CatNotFound", {
-  id: Schema.Number
+  id: Schema.Number // Refers to the ID by which cat was not found
 }) {}
 
 export class CatsApiGroup extends HttpApiGroup.make("cats")
@@ -25,7 +14,7 @@ export class CatsApiGroup extends HttpApiGroup.make("cats")
     HttpApiEndpoint.get("getCatById", "/cats/:id")
       .addSuccess(Cat)
       .addError(CatNotFound, { status: 404 })
-      .setPath(Schema.Struct({ id: CatIdFromString })) // Corrected to use CatIdFromString
+      .setPath(Schema.Struct({ id: CatIdFromString }))
   )
   .add(
     HttpApiEndpoint.post("createCat", "/cats")
@@ -37,21 +26,21 @@ export class CatsApiGroup extends HttpApiGroup.make("cats")
       }))
   )
   .add(
-    HttpApiEndpoint.patch("updateCat", "/cats/:id") // Changed from "completeTodo" to "updateCat"
+    HttpApiEndpoint.patch("updateCat", "/cats/:id")
       .addSuccess(Cat)
       .addError(CatNotFound, { status: 404 })
-      .setPath(Schema.Struct({ id: CatIdFromString })) // Corrected to use CatIdFromString
-      .setPayload(Schema.Struct({ // Added payload for update
+      .setPath(Schema.Struct({ id: CatIdFromString }))
+      .setPayload(Schema.Struct({
         name: Schema.OptionFromNullOr(Schema.NonEmptyTrimmedString),
         breed: Schema.OptionFromNullOr(Schema.NonEmptyTrimmedString),
         age: Schema.OptionFromNullOr(Schema.Number)
       }))
   )
   .add(
-    HttpApiEndpoint.del("deleteCat", "/cats/:id") // Changed from "removeTodo" to "deleteCat"
-      .addSuccess(Schema.Void) // Success is Void as nothing is returned
+    HttpApiEndpoint.del("deleteCat", "/cats/:id")
+      .addSuccess(Schema.Void)
       .addError(CatNotFound, { status: 404 })
-      .setPath(Schema.Struct({ id: CatIdFromString })) // Corrected to use CatIdFromString
+      .setPath(Schema.Struct({ id: CatIdFromString }))
   )
 {}
